initial commit
Rock Paper Scissor is basically a game strategy and it is played using hands. So i have developed this game in my computer using Python where the code needs to take the user choice and compare it with the computers choice(using random module in python). Every time the user or computer wins the score would be increased by 1, and the total no of plays would be 3.